dashboard-service/requirements.txt
# dashboard_service/requirements.txt
Flask
requests
firebase-admin
pillow